Remove inspector.js
